requirements.txt
torch
torchvision
flask
python-dotenv
pillow
numpy
matplotlib
scikit-learn
segmentation-models-pytorch
opencv-python